Refactorizacion

+++ src/main/java/edu/eci/pdsw/sampleprj/dao/mybatis/MyBATISClienteDAO.java
@@ -44,7 +44,7 @@ public class MyBATISClienteDAO implements ClienteDAO{
     @Override
     public void addItemACliente(long docu, Item i, Date date,int numdias) throws PersistenceException {
         try{
-            clienteMapper.agregarItemRentadoACliente(docu, i.getId(), date, new Date(date.getTime()+(long)numdias*MILLISECONDS_IN_DAY));
+            clienteMapper.agregarItemRentadoACliente(docu, i.getId(), date, new Date(date.getTime()+numdias*MILLISECONDS_IN_DAY));
         }catch(org.apache.ibatis.exceptions.PersistenceException e){
             throw new PersistenceException("Error al agregar el item con id: "+i.getId()+", al cliente con documento: "+docu,e);
         }
